Move transcription line to experimental/ Move the transcription line components and models to drawingTools/experimental. Add Tool to the tools exports, and give it a type.
