+++ packages/lib-classifier/src/plugins/drawingTools/components/index.js
@@ -5,4 +5,5 @@ export { default as Mark } from './Mark'
 export { default as Line } from './Line'
 export { default as Point } from './Point'
 export { default as Rectangle } from './Rectangle'
-export { default as TranscriptionLine } from './TranscriptionLine'
+export { default as TranscriptionLine } from '@plugins/drawingTools/experimental/components/TranscriptionLine'
+
