+++ packages/lib-classifier/src/plugins/drawingTools/experimental/components/TranscriptionLine/index.js
