+++ packages/lib-classifier/src/plugins/drawingTools/experimental/components/index.js
@@ -1,0 +1,1 @@
+export { default as TranscriptionLine } from './TranscriptionLine'
