+++ packages/lib-classifier/src/plugins/drawingTools/experimental/models/tools/TranscriptionLineTool/index.js
