+++ packages/lib-classifier/src/plugins/drawingTools/experimental/models/tools/index.js
@@ -1,0 +1,1 @@
+export { default as TranscriptionLineTool } from './TranscriptionLineTool'
