+++ packages/lib-classifier/src/plugins/drawingTools/models/marks/index.js
@@ -2,4 +2,5 @@ export { default as Line } from './Line'
 export { default as Mark } from './Mark'
 export { default as Point } from './Point'
 export { default as Rectangle } from './Rectangle'
-export { default as TranscriptionLine } from './TranscriptionLine'
+export { default as TranscriptionLine } from '@plugins/drawingTools/experimental/models/marks/TranscriptionLine'
+
