+++ packages/lib-classifier/src/plugins/drawingTools/models/tools/index.js
@@ -2,5 +2,5 @@ export { default as LineTool } from './LineTool'
 export { default as PointTool } from './PointTool'
 export { default as RectangleTool } from './RectangleTool'
 export { default as Tool } from './Tool'
-export { default as TranscriptionLineTool } from './TranscriptionLineTool'
+export { default as TranscriptionLineTool } from '@plugins/drawingTools/experimental/models/tools/TranscriptionLineTool'
